requirements.txt
SpeechRecognition
requests
pyaudio
google-genai
edge-tts
asyncio
pygame
pyautogui
colorama
psutil
opencv-python
Pillow
duckduckgo_search
geopy
thefuzz
pyperclip
python-dotenv